+++ test/kde.jl
@@ -207,27 +207,31 @@ end
 end
 
 @testset "Histogramming Accuracy" begin
-    # This range is an example taken from Julia's test suite for its range handling,
-    # called out as having challenging values
-    r = 1.0:1/49:27.0
-    v = Vector(r)
+    r64 = 2e0:1e0/49:28e0
+    r32 = 2f0:1f0/49:28f0
+    l64 = LinRange(r64[1], r64[end], length(r64))
+    l32 = LinRange(r32[1], r32[end], length(r32))
 
-    # For regular histogram binning, using the bin edges as values must result in a uniform
-    # distribution except the last bin which is doubled (due to being closed on the right).
-    ν, H = KDE._kdebin(KDE.HistogramBinning(), v, nothing, first(r), last(r), length(r) - 1)
-    @test ν == length(r)
-    @test all(@view(H[1:end-1]) .== H[1])
-    @test H[end] == 2H[1]
+    @testset "$r" for r in (r64, l64, r32, l32)
+        v = Vector(r)
 
-    # For linear binning, the first and last bins differ from the rest, getting all of the
-    # weight from the two edges but also gaining half a contribution from their (only)
-    # neighbors. (The remaining interior bins give up half of their weight but
-    # simultaneously gain from a neighbor, so they are unchanged.)
-    ν, H = KDE._kdebin(KDE.LinearBinning(), v, nothing, first(r), last(r), length(r) - 1)
-    @test ν == length(r)
-    @test all(@view(H[2:end-1]) .≈ H[2])
-    @test H[end] ≈ H[1]
-    @test H[1] ≈ 1.5H[2]
+        # For regular histogram binning, using the bin edges as values must result in a uniform
+        # distribution except the last bin which is doubled (due to being closed on the right).
+        ν, H = KDE._kdebin(KDE.HistogramBinning(), v, nothing, r[1], r[end], length(r) - 1)
+        @test ν == length(r)
+        @test_broken all(@view(H[1:end-1]) .== H[1])
+        @test H[end] == 2H[1]
+
+        # For linear binning, the first and last bins differ from the rest, getting all of the
+        # weight from the two edges but also gaining half a contribution from their (only)
+        # neighbors. (The remaining interior bins give up half of their weight but
+        # simultaneously gain from a neighbor, so they are unchanged.)
+        ν, H = KDE._kdebin(KDE.LinearBinning(), v, nothing, r[1], r[end], length(r) - 1)
+        @test ν == length(r)
+        @test all(@view(H[2:end-1]) .≈ H[2])
+        @test H[end] ≈ H[1]
+        @test H[1] ≈ 1.5H[2]
+    end
 
     # A case where naively calculating the cell index and weight factors suffers from the
     # limits of finite floating point calculations, e.g.